Basic working POM for dev_4_4
The version of this pom is 4.4.1, meaning the first
release of the pom for the 4.4 branch. Each modification
to the pom should bump the version.

This is currently hard-coded to a SNAPSHOT release, but
as soon as possible, 4.4 artifacts should be uploaded
to artifactory in order to move to proper release versions.